Move the test script
